refactor: extract app sub-modules, remove dead code, fix glob imports
- Remove #[allow(dead_code)] from ui/mod.rs: gate find_menu_item/find_in_items
  with #[cfg(test)], use name field in log output
- Replace glob import 'use crate::types::*' with explicit imports in app
- Extract spawn_periodic_event helper to deduplicate 3 identical async loops
- Restructure app.rs into app/ module with tray.rs and fs_watcher.rs
- Deduplicate PollWatcher/RecommendedWatcher setup into create_watcher()
